+++ shared-types/src/index.ts
@@ -1,1 +1,4 @@
-export {};
+export * from './domain';
+export * from './dto';
+export * from './dates';
+export * from './events';